Feat: implemnting Dapper's Multi-Mapping Feature
In the Dapper Query method, a Lambda function is used to help with mapping each question. By using Doctionary we can kee ptrack of the questions that have already been created which then allows a new instance list<AnswerGetResponse> for the answers of the new questions. In the generic section I provided three responses. 1st two are for mapping the models and last one is for the end results on what should be mapped to. Dapper has a feature called 'splitOn' which allwos you to provide the property for which is needed to be mapped. In this case everything before QuestionId (which is in AnswerGetResponse) goes to QuestiongetManyResponse and everything after, including 'QuestionId', goes int the AnswerGetResponse model. Distinct removes any duplicated questions.